Refactoring CommandEval arguments.
Was too eager when I earlier made many of the arguments to the relation implicit.  Since Imp is imperative and has mutable state, the definition of equivalence between programs often comes down to 'produces the same global state'.  Accordingly, several theorems require explicit manipulation states, which was painful when they were implicit arguments.

Refactored to make the state arguments explicit but left the other non-hypothesis arguments implicit.  Also, moved the state arguments to immediately follow the implicit arguments for consistency.  Now, if you have to use the '@' syntax to include all arguments, you only have to inject fields at the beginning, not in the middle, of the pattern.